example_django_site/urls.py
"""example_django_site URL Configuration

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/1.11/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  url(r'^$', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  url(r'^$', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.conf.urls import url, include
    2. Add a URL to urlpatterns:  url(r'^blog/', include('blog.urls'))
"""
from django.conf.urls import url, include
from django.contrib import admin
from rest_framework.authtoken import views
from rest_framework_swagger.views import get_swagger_view
from django.shortcuts import redirect


def redirect_to_docs(request):
    return redirect('docs:main_doc_view')

urlpatterns = [
    url(r'^admin/', admin.site.urls),
    url('^api/', include('api.urls', namespace='api')),
    url('^docs/', include('docs.urls', namespace='docs')),
    url(r'^api-auth/', include('rest_framework.urls')),
    url(r'^api-token-auth/', views.obtain_auth_token),
    url(r'swagger/', get_swagger_view()),
    url('^$', redirect_to_docs),
]
