requirements.txt
dj-database-url==0.4.2
Django==1.11.1
django-filter==1.0.3
djangorestframework==3.6.3
gunicorn==19.7.1
Markdown==2.6.8
psycopg2==2.7.1
pytz==2017.2
whitenoise==3.3.0
requests==2.18.1
django-safedelete
django-rest-swagger
coreapi
